*NEW* [SIS-049] Get featured activity cards on home page (#126)
* feat: new version of SSI049 branch, made to bypass build and git issues

* feat: new SIS049 branch, with edits and build issues fixed

duplicated the SIS049 branch to resolve conflicts with git and node_modules. added suggestions,
except for some minor styling issues

* linting issues

* fixed date displaying and styling merge changes, getFeaturedActivityCards now displays from past 2 weeks

+++ src/services/ActivityCardService.js
@@ -7,9 +7,9 @@ import {
 } from './routes/Local';
 import { MAINDIRECTORY } from './constants';
 import { AccessToken } from './models';
+import { Buffer } from 'buffer';
 import axios from 'axios';
 import { DRIVE_API_URLS } from './config.json';
-import configureStore from './configureStore';
 
 const ActivityCardService = {
   // All APIs for ActivityCards should be here
@@ -24,15 +24,17 @@ const ActivityCardService = {
     try {
       //Retrieve the last week from browser, convert to ISO for use in query
       const weekAgo = new Date(
-        Date.now() - 8 * 24 * 60 * 60 * 1000,
+        Date.now() - 15 * 24 * 60 * 60 * 1000,
       ).toISOString();
 
       //Set up GET url, query, and path to Featured Card directory
       var downloadUrl =
         DRIVE_API_URLS.SEARCH_FILES + DRIVE_API_URLS.SEARCH_PARAMETERS;
-      const searchQuery =
-        "modifiedTime >= '" + weekAgo + "' and mimeType contains 'image/'";
       var path = MAINDIRECTORY + '/FeaturedActivityCards/';
+      const searchQuery =
+        "modifiedTime >= '" +
+        weekAgo +
+        "' and mimeType contains 'image/' and fullText contains 'ACTVT'";
       const params = {
         q: searchQuery,
         fields:
@@ -41,7 +43,7 @@ const ActivityCardService = {
 
       //Retreive all Drive files meeting the params
       const response = await axios.get(downloadUrl, { params }).catch(error => {
-        console.error('ERROR IN GETTING FEATURED ACTIVITY CARDS: ' + error);
+        console.log('ERROR IN GETTING FEATURED ACTIVITY CARDS: ' + error);
       });
 
       //Access the Array of all files and set up path Array (to be returned)
@@ -49,26 +51,30 @@ const ActivityCardService = {
       const files_list = driveFiles.files;
       var pathArr = [];
 
+      if(files_list.length == 0){
+        return [];
+      }
+
       //Delete anything that may currently be in the Featured Cards directory, make the new path with no contents
-      if (files_list.length != 0) {
+      if (files_list.length != 0 && (await checkFileExists(path))) {
         await deleteFile(path);
         await makeDirectory(path);
-      } else {
-        return;
       }
+
 
       //if new cards were found, save them into the empty directory path
       for (var i = 0; i < files_list.length; i++) {
         await this.downloadActivityCard(files_list[i].id);
 
         //once downloaded, check if the file exists. If it does, add the name to a .txt file, and add the path to pathArr
-        if (await checkFileExists(path + '/' + files_list[i].id + '/')) {
+        if (await checkFileExists(path + files_list[i].id + '/')) {
+          const cardName = files_list[i].name;
           await writeFile(
             false,
-            path + '/' + files_list[i].id + '/' + 'cardName.txt',
-            files_list[i].name,
+            path + files_list[i].id + '/' + 'cardName.txt',
+            cardName + '\n',
           );
-          pathArr.push(path + '/' + files_list[i].id + '/');
+          pathArr.push(path + files_list[i].id + '/');
         } else {
           throw new Error(
             'Error occurred in downloading a Featured Activity Card.',
@@ -80,6 +86,7 @@ const ActivityCardService = {
     } catch (e) {
       // There was an error, catch it and do something with it
       console.log('ERROR IN LISTING FEATURED ACTIVITY CARDS: ' + e);
+      return [];
     }
   },
 
@@ -92,13 +99,11 @@ const ActivityCardService = {
   downloadActivityCard: async function (id) {
     try {
       const params = {
-        fields:
-          'kind,driveId,mimeType,id,name,teamDriveId,thumbnailLink,webContentLink,modifiedTime',
         alt: 'media',
+        responseType: 'arraybuffer',
       };
-      const dirPath = MAINDIRECTORY + '/FeaturedActivityCards';
-      const filePath = `${dirPath}/${id}.jpg`;
-      let card = {};
+      const dirPath = MAINDIRECTORY + '/FeaturedActivityCards/' + id;
+      const filePath = `${dirPath}/cardImage.jpg`;
 
       //check if file exists
       if (await checkFileExists(filePath)) {
@@ -107,20 +112,21 @@ const ActivityCardService = {
 
       //set up the get URL, then call axios for response
       const downloadUrl =
-        DRIVE_API_URLS.SEARCH_FILES + id + DRIVE_API_URLS.SEARCH_PARAMETERS;
-      console.log('DOWNLOAD url' + downloadUrl);
-      console.log(id);
+        DRIVE_API_URLS.SEARCH_FILES +
+        '/' +
+        id +
+        DRIVE_API_URLS.SEARCH_PARAMETERS;
 
-      const response = await axios.get(downloadUrl, { params }).catch(error => {
-        console.error('ERROR IN DOWNLOADING ACTIVITY CARD: ' + error);
-      });
-
-      console.log(response);
-      card = response.data;
-
-      //make directory for the newly downloaded card, write the card into this path and return
+      //get the response as an arrayBuffer to be read by RNFS
+      const response = await axios
+        .get(downloadUrl, { params: params, responseType: 'arraybuffer' })
+        .catch(error => {
+          console.log('ERROR IN DOWNLOADING ACTIVITY CARD: ' + error);
+        });
+      
       await makeDirectory(dirPath);
-      await writeFile(true, filePath, card);
+      await writeFile(true, filePath, Buffer.from(response.data, 'base64'));
+
       return filePath;
     } catch (e) {
       console.log('ERROR IN DOWNLOADING ACTIVITY CARD: ' + e);
